THere is a nasty segfault somewhere blocking up most of the go tests

+++ go/falcon-core/generic/farraydouble/fArrayDouble_test.go
@@ -90,7 +90,7 @@ func TestFArrayDouble_ShapeDimension(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Dimension failed: %v", err)
 	}
-	if dim != len(defaultShape) {
+	if dim != uint32(len(defaultShape)) {
 		t.Errorf("Expected dimension %d, got %d", len(defaultShape), dim)
 	}
 }
@@ -585,7 +585,7 @@ func TestFArrayDouble_RemoveOffsetSum(t *testing.T) {
 func TestFArrayDouble_ReshapeWhereFlip(t *testing.T) {
 	a, _ := FromData(defaultData, defaultShape)
 	defer a.Close()
-	_, err := a.Reshape([]int{4, 1})
+	_, err := a.Reshape([]int32{4, 1})
 	if err != nil {
 		t.Errorf("Reshape failed: %v", err)
 	}
